+++ lab_05/src/time_func.c
@@ -1,6 +1,6 @@
 #include "../inc/time_func.h"
 
-double rtime(int left, int right)
+double rtime(double left, double right)
 {
     double a = (right - left) * ((double)rand() / (double)RAND_MAX) + left;
     return (double)((int)(a * 1000)) / 1000;
